fix(sweep): address Thermo code-quality re-review findings
Remove invalid workflow guards, split codex stream handling, tighten audit
typing, drop dead exports, and wire npm lock discipline into CI/lint.
