+++ .github/workflows/integration.yml
@@ -61,7 +61,9 @@ jobs:
         with:
           node-version: "22"
           cache: npm
-          cache-dependency-path: docker/agent-clis/package-lock.json
+          cache-dependency-path: |
+            docker/agent-clis/package-lock.json
+            tools/package-lock.json
       - name: npm audit (agent-clis lockfile)
         run: make npm-audit
       - name: actionlint + zizmor
