+++ src/mergecraft/mcp/review.py
@@ -409,9 +409,9 @@ def create_pull_request_review_tool(ctx: ToolContext):
             submission_payload = _legacy_params_to_submission(params)
             if submission_payload is not None:
                 recorded = record_validated_terminal_submission(ctx, submission_payload)
-                from mergecraft.mcp.terminal_hooks import after_terminal_submission_recorded
+                from mergecraft.enterprise.audit import maybe_audit_blocking_terminal_submission
 
-                after_terminal_submission_recorded(ctx, recorded)
+                maybe_audit_blocking_terminal_submission(ctx, recorded)
         else:
             _reject_mismatched_publication(ctx.tool_state.terminal_submission, params)
             revalidate_recorded_submission(ctx)
